requirements.txt
streamlit>=1.40
google-generativeai>=0.8
psycopg[binary]>=3.2
pgvector>=0.3
python-dotenv>=1.0
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.0
pypdf>=5.0
